Fix CdcIterator serializer to write byte not short when rangeFilter is null

+++ cdc/src/main/java/org/apache/cassandra/spark/cdc/jdk/JdkCdcIterator.java
@@ -479,7 +479,7 @@ public abstract class JdkCdcIterator implements AutoCloseable, IStreamScanner<Jd
             }
             else
             {
-                out.writeShort(-1);
+                out.writeByte(-1);
             }
 
             // write start markers
